requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.11.0
    # via
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
build==1.4.0
    # via pip-tools
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
cryptography==46.0.3
    # via oracledb
django==5.2.7
    # via
    #   -r requirements.in
    #   django-computedfields
    #   django-cors-headers
    #   django-extended-ol
    #   django-fast-update
    #   django-filter
    #   django-geojson
    #   django-revproxy
    #   djangorestframework
    #   djangorestframework-gis
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-computedfields==0.3.5
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-extended-ol==1.2.1
    # via -r requirements.in
django-fast-update==0.3.0
    # via
    #   -r requirements.in
    #   django-computedfields
django-filter==25.2
    # via djangorestframework-gis
django-geojson==4.2.0
    # via -r requirements.in
django-revproxy==0.13.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-gis
    #   drf-spectacular
djangorestframework-gis==1.2.0
    # via -r requirements.in
drf-spectacular==0.29.0
    # via -r requirements.in
drf-spectacular-sidecar==2026.1.1
    # via -r requirements.in
et-xmlfile==2.0.0
    # via openpyxl
idna==3.11
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
openpyxl==3.1.5
    # via -r requirements.in
oracledb==3.4.1
    # via -r requirements.in
packaging==25.0
    # via build
phonenumbers==9.0.22
    # via -r requirements.in
pip-tools==7.5.2
    # via -r requirements.in
psycopg[binary]==3.3.2
    # via -r requirements.in
psycopg-binary==3.3.2
    # via psycopg
pycparser==2.23
    # via cffi
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.2.1
    # via -r requirements.in
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via -r requirements.in
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via
    #   django-computedfields
    #   oracledb
tzdata==2025.3
    # via
    #   django
    #   psycopg
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.6.3
    # via
    #   django-revproxy
    #   requests
wheel==0.45.1
    # via pip-tools
whitenoise==6.11.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
